requirements.txt
certifi==2017.4.17
chardet==3.0.4
Django==1.11.2
django-oauth-toolkit==1.0.0
djangorestframework==3.6.3
idna==2.5
oauthlib==2.0.2
psycopg2==2.6.2
pytz==2017.2
requests==2.18.1
urllib3==1.21.1
django-rest-swagger